v2.0
Now, rooms do not have a text name. Instead, they have only a list of tags. If one of these tags is in the access list of a person, this person might open the door.
There is a new emergency mode with ASCII art fun.
Also, there are new test cases.
The scopes are finally finished. Using scopes and commands, you can get a lot of information about people, access lists and rooms